requirements.txt
aiobotocore
aiofiles
aiohttp
aioitertools
aiosignal
aiosqlite
alabaster
albumentations
alembic
anaconda-anon-usage
anaconda-catalogs
anaconda-client
anaconda-cloud-auth
anaconda-navigator
anaconda-project
anyio
appdirs
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
asttokens
async-generator
async-timeout
atomicwrites
attrs
Automat
autopep8
Babel
backcall
backports.functools-lru-cache
backports.tempfile
backports.weakref
bcrypt
beautifulsoup4
binaryornot
biopython
black
bleach
blinker
bokeh
boltons
botocore
Bottleneck
brotlipy
certifi
certipy
cffi
chardet
charset-normalizer
click
cloudpickle
clyent
colorama
colorcet
comm
conda
conda-build
conda-content-trust
conda_index
conda-libmamba-solver
conda-pack
conda-package-handling
conda_package_streaming
conda-repo-cli
conda-token
conda-verify
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cytoolz
daal4py
dask
datasets
datashader
datashape
debugpy
decorator
defusedxml
diff-match-patch
dill
distributed
docstring-to-markdown
docutils
entrypoints
et-xmlfile
executing
fastjsonschema
filelock
flake8
Flask
fonttools
frozenlist
fsspec
future
gensim
glob2
gmpy2
greenlet
h5py
HeapDict
holoviews
huggingface-hub
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
importlib-resources
incremental
inflection
iniconfig
intake
intervaltree
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jaraco.classes
jedi
jeepney
jellyfish
Jinja2
jinja2-time
jmespath
joblib
json5
jsonpatch
jsonpointer
jsonschema
jupyter
jupyter_client
jupyter-console
jupyter_core
jupyter-events
jupyter-server
jupyter_server_fileid
jupyter_server_ydoc
jupyter-telemetry
jupyter-ydoc
jupyterhub
jupyterlab
jupyterlab-pygments
jupyterlab_server
jupyterlab-widgets
kaleido
keyring
kiwisolver
lazy_loader
lazy-object-proxy
libarchive-c
libmambapy
lightning-utilities
linkify-it-py
llvmlite
lmdb
locket
lxml
lz4
Mako
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
mistune
mkl-fft
mkl-random
mkl-service
monai
more-itertools
mpmath
msgpack
multidict
multipledispatch
multiprocess
munkres
mypy-extensions
navigator-updater
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numba
numexpr
numpy
numpydoc
oauthlib
openpyxl
packaging
pamela
pandas
pandocfilters
panel
param
parsel
parso
partd
pathlib
pathspec
patsy
pep8
pexpect
pickleshare
Pillow
pip
pkce
pkginfo
platformdirs
plotly
pluggy
ply
poyo
prometheus-client
prompt-toolkit
Protego
psutil
ptyprocess
pure-eval
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pycodestyle
pycosat
pycparser
pyct
pycurl
pydantic
PyDispatcher
pydocstyle
pyerfa
pyflakes
Pygments
PyJWT
pylint
pylint-venv
pyls-spyder
pyodbc
pyOpenSSL
pyparsing
PyQt5
PyQt5-sip
PyQtWebEngine
pyrsistent
PySocks
pytest
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-slugify
python-snappy
pytoolconfig
pytorch-msssim
pytz
pyviz-comms
PyWavelets
pyxdg
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
qudida
queuelib
regex
requests
requests-file
requests-toolbelt
responses
rfc3339-validator
rfc3986-validator
rope
Rtree
ruamel.yaml
ruamel-yaml-conda
s3fs
safetensors
scikit-image
scikit-learn
scikit-learn-intelex
scipy
Scrapy
seaborn
SecretStorage
Send2Trash
service-identity
setuptools
SimpleITK
sip
six
smart-open
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
stack-data
statsmodels
sudospawner
sympy
tables
tabulate
TBB
tblib
tenacity
terminado
text-unidecode
textdistance
threadpoolctl
three-merge
tifffile
tinycss2
tldextract
tokenizers
toml
tomlkit
toolz
torchmetrics
torchsummary
tornado
tqdm
traitlets
transformers
Twisted
typing_extensions
tzdata
uc-micro-py
ujson
Unidecode
urllib3
w3lib
watchdog
wcwidth
webencodings
websocket-client
Werkzeug
whatthepatch
wheel
widgetsnbextension
wrapt
wurlitzer
xarray
xxhash
xyzservices
y-py
yapf
yarl
ypy-websocket
zict
zipp
zope.interface
zstandard
